move enums
